+++ frontend/.eslintrc.cjs
@@ -126,7 +126,7 @@ module.exports = {
       }
     },
     {
-      files: ['src/domain/types/*.ts', 'src/domain/**/types.ts', 'src/domain/shared_slices/**/*.ts', 'src/**/slice.ts'],
+      files: ['src/domain/schemas/*.ts', 'src/domain/**/schemas.ts', 'src/domain/shared_slices/**/*.ts', 'src/**/slice.ts'],
       plugins: ['no-null'],
       rules: {
         'no-param-reassign': 'off'
