+++ frontend/cypress/e2e/side_window/prior_notification_list/filter_bar.spec.ts
@@ -227,7 +227,7 @@ context('Side Window > Prior Notification List > VesselFilter Bar', () => {
     cy.get('.Table-SimpleTable tr').should('have.length.to.be.greaterThan', 0)
   })
 
-  it('Should filter prior notifications by types', () => {
+  it('Should filter prior notifications by schemas', () => {
     openSideWindowPriorNotificationListAsSuperUser()
 
     cy.intercept('GET', `${apiPathBase}*priorNotificationTypes=${encodeURI('Préavis type A,Préavis type C')}*`).as(
