+++ frontend/src/env.d.ts
@@ -1,5 +1,4 @@
 // https://vitejs.dev/guide/env-and-mode#intellisense-for-typescript
-/// <reference types="vite/client" />
 
 interface ImportMetaEnv {
   readonly FRONTEND_GEOSERVER_LOCAL_URL: string
