+++ frontend/src/features/Vessel/vesselNavApi.ts
@@ -1,4 +1,5 @@
 import { monitorfishLightApi } from '@api/api'
+import { VesselLastPositionSchema } from '@features/Vessel/schemas/VesselLastPositionSchema'
 import { Vessel } from '@features/Vessel/Vessel.types'
 
 export const vesselNavApi = monitorfishLightApi.injectEndpoints({
@@ -6,7 +7,7 @@ export const vesselNavApi = monitorfishLightApi.injectEndpoints({
     getVesselsLastPositions: builder.query<Vessel.VesselLastPosition[], void>({
       query: () => `/v1/vessels`,
       transformResponse: (baseQueryReturnValue: Vessel.VesselLastPosition[]) =>
-        baseQueryReturnValue.map(LastPosition => Vessel.VesselLastPositionSchema.parse(LastPosition))
+        baseQueryReturnValue.map(LastPosition => VesselLastPositionSchema.parse(LastPosition))
     })
   })
 })
